VirusReportingService sends results to endpoint
The MetadataListener::Job will take arguments, including an endpoint and
api token, and report the results of the virus scan to the requesting
application.

If no endpoint is provided, the results are logged via STDOUT.

For now, the job will check for viruses but not report the results until
we have decided how we want to do that.