ml executable
Adds an executable that starts a Pry session with all of the
metadata-listener code loaded. This is useful when debugging or running
commands from inside the container.

Typically, pry is reserved for development environments, but because
this is a non-standard kind of application, Pry is really just
functioning as a console with nicer features.