Disable servlet (#40)
Removes the Fits servlet and runs it as a shell command. This
reduces the memory footprint considerably, at the cost of taking
longer to run the job.

When calling Open3, it will throw an exception if the command path
doesn't exist. However, if the command does execute, but is _not_
successful, we want to check for that too, and raise our own error.

Co-authored-by: Adam Wead <awead@users.noreply.github.com>