Call Clamby::Command directly
Because the Clamby gem is supporting legacy behavior, if clamscan
encounters an unknown error, Clamby will indicate that the file has a
virus, when in fact, it may not. This is a false-positive case, and the
legacy behavior takes a slightly paranoid approach. While this is
understandable, it creates problems on our end, and we'd rather know
that file caused an error when scanned and not that it "has a virus."

To get around this, we're calling the same methods that the gem does and
handling the errors in our own way. An exit code of 0 means "no virus",
an exit code of 1 means "definitely virus", and anything else raises an
error and we figure it out later, re-try, etc.